upload_mesh/requirements.txt
# Настройка виртуального окружения Python для Linux
#
# 1. Создание виртуального окружения:
#    python3 -m venv venv
#
# 2. Активация виртуального окружения:
#    source venv/bin/activate
#
# 3. Установка зависимостей из этого файла:
#    pip install -r requirements.txt
#
# 4. Деактивация виртуального окружения:
#    deactivate
#
# После установки зависимостей можно запускать build_script.py
# Скрипт автоматически активирует venv при необходимости.

# platformio>=6.1.18
pyserial>=3.5
esptool==4.6.2
# intelhex>=2.3.0
